+++ backend/data/src/main/kotlin/io/tolgee/configuration/tolgee/TolgeeProperties.kt
@@ -109,6 +109,8 @@ open class TolgeeProperties(
   var postgresAutostart: PostgresAutostartProperties = PostgresAutostartProperties(),
   @DocProperty(hidden = true)
   var sendInBlue: SendInBlueProperties = SendInBlueProperties(),
+  @DocProperty(hidden = true)
+  var mailjet: MailjetProperties = MailjetProperties(),
   open var import: ImportProperties = ImportProperties(),
   var rateLimit: RateLimitProperties = RateLimitProperties(),
   @DocProperty(hidden = true)